+++ exporters/otlp/src/otlp_grpc_exporter_options.cc
@@ -29,6 +29,11 @@ OtlpGrpcExporterOptions::OtlpGrpcExporterOptions()
   timeout    = GetOtlpDefaultTracesTimeout();
   metadata   = GetOtlpDefaultTracesHeaders();
   user_agent = GetOtlpDefaultUserAgent();
+
+  max_threads = 0;
+#ifdef ENABLE_ASYNC_EXPORT
+  max_concurrent_requests = 64;
+#endif
 }
 
 OtlpGrpcExporterOptions::~OtlpGrpcExporterOptions() {}