+++ exporters/otlp/src/otlp_grpc_log_record_exporter_options.cc
@@ -27,6 +27,11 @@ OtlpGrpcLogRecordExporterOptions::OtlpGrpcLogRecordExporterOptions()
   timeout    = GetOtlpDefaultLogsTimeout();
   metadata   = GetOtlpDefaultLogsHeaders();
   user_agent = GetOtlpDefaultUserAgent();
+
+  max_threads = 0;
+#ifdef ENABLE_ASYNC_EXPORT
+  max_concurrent_requests = 64;
+#endif
 }
 
 OtlpGrpcLogRecordExporterOptions::~OtlpGrpcLogRecordExporterOptions() {}